+++ src/client/instruction.rs
@@ -3,12 +3,17 @@
 
 //! Instructions that can be sent to the client actor from client handles.
 
+use chrono::TimeDelta;
 use tokio::sync::{broadcast, oneshot};
 
 use crate::{
     Error,
-    proto::common::microgrid::electrical_components::{
-        ElectricalComponent, ElectricalComponentConnection, ElectricalComponentTelemetry,
+    proto::common::{
+        metrics::{Bounds, Metric},
+        microgrid::electrical_components::{
+            ElectricalComponent, ElectricalComponentCategory, ElectricalComponentConnection,
+            ElectricalComponentTelemetry,
+        },
     },
 };
 
@@ -21,12 +26,19 @@ pub(super) enum Instruction {
     },
     ListElectricalComponents {
         electrical_component_ids: Vec<u64>,
-        electrical_component_categories: Vec<i32>,
+        electrical_component_categories: Vec<ElectricalComponentCategory>,
         response_tx: oneshot::Sender<Result<Vec<ElectricalComponent>, Error>>,
     },
     ListElectricalComponentConnections {
         source_electrical_component_ids: Vec<u64>,
         destination_electrical_component_ids: Vec<u64>,
         response_tx: oneshot::Sender<Result<Vec<ElectricalComponentConnection>, Error>>,
+    },
+    AugmentElectricalComponentBounds {
+        electrical_component_id: u64,
+        target_metric: Metric,
+        bounds: Vec<Bounds>,
+        request_lifetime: Option<TimeDelta>,
+        response_tx: oneshot::Sender<Result<Option<chrono::DateTime<chrono::Utc>>, Error>>,
     },
 }