+++ src/client/test_utils.rs
@@ -24,6 +24,7 @@ use crate::{
         },
         google::protobuf,
         microgrid::{
+            AugmentElectricalComponentBoundsRequest, AugmentElectricalComponentBoundsResponse,
             ListElectricalComponentConnectionsRequest, ListElectricalComponentConnectionsResponse,
             ListElectricalComponentsRequest, ListElectricalComponentsResponse,
             ReceiveElectricalComponentTelemetryStreamRequest,
@@ -280,10 +281,20 @@ impl MicrogridApiClient for MockMicrogridApiClient {
         &mut self,
         _request: impl tonic::IntoRequest<ListElectricalComponentsRequest> + Send,
     ) -> std::result::Result<tonic::Response<ListElectricalComponentsResponse>, tonic::Status> {
+        let ListElectricalComponentsRequest {
+            electrical_component_ids,
+            electrical_component_categories,
+        } = _request.into_request().into_inner();
         Ok(Response::new(ListElectricalComponentsResponse {
             electrical_components: self
                 .components
                 .iter()
+                .filter(|c| {
+                    (electrical_component_ids.is_empty()
+                        || electrical_component_ids.contains(&c.component.id))
+                        && (electrical_component_categories.is_empty()
+                            || electrical_component_categories.contains(&c.component.category))
+                })
                 .map(|c| c.component.clone())
                 .collect(),
         }))
@@ -430,6 +441,14 @@ impl MicrogridApiClient for MockMicrogridApiClient {
 
         let stream = tokio_stream::wrappers::ReceiverStream::new(rx);
         Ok(Response::new(stream))
+    }
+
+    async fn augment_electrical_component_bounds(
+        &mut self,
+        _request: impl tonic::IntoRequest<AugmentElectricalComponentBoundsRequest> + Send,
+    ) -> std::result::Result<tonic::Response<AugmentElectricalComponentBoundsResponse>, tonic::Status>
+    {
+        unimplemented!()
     }
 }