+++ src/logical_meter/logical_meter_handle.rs
@@ -6,6 +6,7 @@ use crate::logical_meter::formula::graph_formula_provider::GraphFormulaProvider;
 use crate::{
     client::MicrogridClientHandle,
     error::Error,
+    metric,
     proto::common::microgrid::electrical_components::{
         ElectricalComponent, ElectricalComponentConnection,
     },
@@ -60,7 +61,7 @@ impl LogicalMeterHandle {
 
     /// Returns a receiver that streams samples for the given `metric` at the grid
     /// connection point.
-    pub fn grid<M: super::metric::Metric>(
+    pub fn grid<M: metric::Metric>(
         &mut self,
         metric: M,
     ) -> Result<Formula<M::QuantityType>, Error> {
@@ -75,7 +76,7 @@ impl LogicalMeterHandle {
     /// given battery IDs.
     ///
     /// When `component_ids` is `None`, all batteries in the microgrid are used.
-    pub fn battery<M: super::metric::Metric>(
+    pub fn battery<M: metric::Metric>(
         &mut self,
         component_ids: Option<BTreeSet<u64>>,
         metric: M,
@@ -92,7 +93,7 @@ impl LogicalMeterHandle {
     /// given CHP IDs.
     ///
     /// When `component_ids` is `None`, all CHPs in the microgrid are used.
-    pub fn chp<M: super::metric::Metric>(
+    pub fn chp<M: metric::Metric>(
         &mut self,
         component_ids: Option<BTreeSet<u64>>,
         metric: M,
@@ -109,7 +110,7 @@ impl LogicalMeterHandle {
     /// given PV IDs.
     ///
     /// When `component_ids` is `None`, all PVs in the microgrid are used.
-    pub fn pv<M: super::metric::Metric>(
+    pub fn pv<M: metric::Metric>(
         &mut self,
         component_ids: Option<BTreeSet<u64>>,
         metric: M,
@@ -127,7 +128,7 @@ impl LogicalMeterHandle {
     ///
     /// When `component_ids` is `None`, all EV chargers in the microgrid are
     /// used.
-    pub fn ev_charger<M: super::metric::Metric>(
+    pub fn ev_charger<M: metric::Metric>(
         &mut self,
         component_ids: Option<BTreeSet<u64>>,
         metric: M,
@@ -142,7 +143,7 @@ impl LogicalMeterHandle {
 
     /// Returns a receiver that streams samples for the given `metric` for the
     /// logical `consumer` in the microgrid.
-    pub fn consumer<M: super::metric::Metric>(
+    pub fn consumer<M: metric::Metric>(
         &mut self,
         metric: M,
     ) -> Result<Formula<M::QuantityType>, Error> {
@@ -155,7 +156,7 @@ impl LogicalMeterHandle {
 
     /// Returns a receiver that streams samples for the given `metric` for the
     /// logical `producer` in the microgrid.
-    pub fn producer<M: super::metric::Metric>(
+    pub fn producer<M: metric::Metric>(
         &mut self,
         metric: M,
     ) -> Result<Formula<M::QuantityType>, Error> {
@@ -168,7 +169,7 @@ impl LogicalMeterHandle {
 
     /// Returns a receiver that streams samples for the given `metric` for the
     /// given component ID.
-    pub fn component<M: super::metric::Metric>(
+    pub fn component<M: metric::Metric>(
         &mut self,
         component_id: u64,
         metric: M,