+++ src/sample.rs
@@ -5,12 +5,26 @@ use chrono::{DateTime, Utc};
 
 /// Represents a measurement of a microgrid metric, made at a specific time.
 #[derive(Copy, Clone, Debug, Default)]
-pub struct Sample<Q: Copy + Clone + std::fmt::Debug + Default> {
+pub struct Sample<Q: Copy + Clone + std::fmt::Debug + Default + std::fmt::Display> {
     pub(crate) timestamp: DateTime<Utc>,
     pub(crate) value: Option<Q>,
 }
 
-impl<Q: Copy + Clone + Default + std::fmt::Debug> frequenz_resampling::Sample for Sample<Q> {
+impl<Q: crate::quantity::Quantity> std::fmt::Display for Sample<Q> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Sample({}, ", self.timestamp)?;
+
+        if let Some(value) = self.value {
+            write!(f, " {})", value)
+        } else {
+            write!(f, " None)")
+        }
+    }
+}
+
+impl<Q: Copy + Clone + Default + std::fmt::Debug + std::fmt::Display> frequenz_resampling::Sample
+    for Sample<Q>
+{
     type Value = Q;
 
     fn new(timestamp: DateTime<Utc>, value: Option<Self::Value>) -> Self {
@@ -26,7 +40,7 @@ impl<Q: Copy + Clone + Default + std::fmt::Debug> frequenz_resampling::Sample fo
     }
 }
 
-impl<Q: Copy + Clone + Default + std::fmt::Debug> Sample<Q> {
+impl<Q: Copy + Clone + Default + std::fmt::Debug + std::fmt::Display> Sample<Q> {
     /// Creates a new `Sample` with the given timestamp and value.
     pub fn new(timestamp: DateTime<Utc>, value: Option<Q>) -> Self {
         Self { timestamp, value }